update main.c to read files and stdout

+++ compiler/lexical_analysis/c/run.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 
-gcc main.c lexer.c -o lexer
+gcc main.c -o lexer
 ./lexer testing.lua
 echo $?
+if [ "$1" == "--no-rm" ]; then
+    exit 0
+fi
 rm ./lexer
